refactor plots to new script
my problem with a prior (reverted) plot script was not realizing ggproto objects can be passed jointly through lists but may not be added to one another. even recursive lists can be passed to ggplot() objects as long as they contain only ggproto objects. This works! This adds option to color by Sex as well.